#90 refactor: product, activity 타입 분리 및 수정
