+++ src/components/Profile/ProductList.tsx
@@ -1,26 +1,15 @@
 import { Loading } from '@/shared/ui/Icon';
 import ProductCard from '@/components/@common/ProductCard';
+import { ProductTypes } from '@/components/Profile/types/productType';
 
-export interface UserProductData {
-  updatedAt: string;
-  createdAt: string;
-  writerId: number;
-  categoryId: number;
-  favoriteCount: number;
-  reviewCount: number;
-  rating: number;
-  image: string;
-  name: string;
-  id: number;
-}
-export interface ProductListProps {
-  productsList: UserProductData[];
+interface ProductListProps {
+  productsList: ProductTypes[];
   content: string;
 }
 export const ProductList = ({ productsList, content }: ProductListProps) => {
   return productsList?.length ? (
     <div className="grid lg:grid-cols-3 md:grid-cols-2 mobile:grid-cols-2 gap-[20px] ">
-      {productsList.map((product: UserProductData) => (
+      {productsList.map((product: ProductTypes) => (
         <ProductCard key={product?.id} product={product} />
       ))}
     </div>
