+++ src/components/Profile/types/productType.ts
@@ -1,0 +1,20 @@
+import { PRODUCT_MENU } from '../constants/PRODUCT_MENU';
+
+export interface ProductTypes {
+  updatedAt: string;
+  createdAt: string;
+  writerId: number;
+  categoryId: number;
+  favoriteCount: number;
+  reviewCount: number;
+  rating: number;
+  image: string;
+  name: string;
+  id: number;
+}
+
+export interface ProductCardProps {
+  product: ProductTypes;
+}
+
+export type ProductMenuType = (typeof PRODUCT_MENU)[number];
