+++ src/components/Profile/types/userActivityType.ts
@@ -1,0 +1,6 @@
+import { IconType } from '@/shared/ui/Icon/types/iconType';
+
+export interface ActivityCardProps {
+  title: string;
+  icon?: IconType;
+}
